댓글 디자인 완료.맵 제작 시작

+++ src/main/webapp/WEB-INF/views/boardRead.jsp
@@ -49,12 +49,15 @@
     <div id="board_content">
       <div>${board.content}</div>
     </div>
+    <h1 id = "comment_h1">전체 댓글 <span>${board.comment_cnt}</span></h1>
     <div id="commentContainer"></div>
-    <h1>--------------------------------------------------------------------------------</h1>
     <div id="replyBox" style="display: none">
-      <input id="replyInput" name="replyInput">
+      <div id="reply_div">
+      <textarea id="replyInput" name="replyInput"></textarea>
       <button id="replySendBtn">저장</button>
-    </div>
+      <button id="reply_cancel_btn">취소</button>
+      </div>
+      </div>
     <div id="comment_send">
       <textarea name="comment" id="comment"></textarea>
       <button id="sendBtn" type="button">SEND</button>
@@ -75,6 +78,11 @@
     showList(bno);
     includeError();
     removeError();
+    $('#remove_a').click(function(event) {
+      if(!confirm('정말로 삭제하시겠습니까?')) {
+        event.preventDefault();
+      }
+    });
   })
   // ********************************************************************************************************************************
   // error 메세지 다루는 함수
@@ -128,7 +136,6 @@
         $("#comment").val('');
       },
       error   : function(data){
-        // alert(data.responseJSON.res);
         $("#error").css("display","flex");
         $("#error p:first-child").html(data.responseJSON.res);
         showList(bno);
@@ -141,11 +148,13 @@
   $("#commentContainer").on("click",".updateBtn",function (){
     // 댓글 수정 버튼이 달려있는 해당 댓글의 li 요소를 매개변수로 넘겨준다.
     let liElement = $(this).closest("li");
+
     makeUpdate(liElement);
   })
   let makeUpdate = function (liElement){
     // 댓글 내용을 가져온다
-    let commentText = liElement.find('span').text();
+    let commentText = liElement.find('span').html();
+    commentText = commentText.replace(/<br>/g, "\n");
     // Check if updateInput and saveBtn already exists, if yes, remove them
     if(liElement.find('.updateInput').length > 0){
       liElement.find('.updateInput').remove();
@@ -153,7 +162,7 @@
     }
     let allDescendants = liElement.find("*");
     allDescendants.css("display","none");
-    liElement.append('<textarea class="updateInput">'+commentText+'</textarea><button class="saveBtn">저장</button><button class="comment_cancel">취소</button>');
+    liElement.append('<div id="update_div"><textarea class="updateInput">'+commentText+'</textarea><button class="saveBtn">저장</button><button class="comment_cancel">취소</button></div>');
     // 여기서 해야할게...기존 댓글의 display 속성을 none으로 만드는것.그리고 취소 버튼도 만들고,거기서 클릭 이벤트를 거는것
   }
   $("#commentContainer").on("click",".saveBtn",function (event){
@@ -161,6 +170,7 @@
     let liElement = $(this).closest("li");
 
     let cno = liElement.attr("data-cno");
+    console.log(cno);
 
     let updatedComment = liElement.find('.updateInput').val();
     update(cno, updatedComment);
@@ -197,19 +207,24 @@
   // ************************************************************************
   // 댓글 삭제 기능
   $("#commentContainer").on("click",".deleteBtn",function (){
-    let cno = $(this).parent().attr("data-cno");
-    let bno = $(this).parent().attr("data-bno");
-    $.ajax({
-      type:'delete',       // 요청 메서드
-      url: '/instorage/comments/'+cno+'?bno='+bno,  // 요청 URI
-      dataType : 'text', // 전송받을 데이터의 타입
-      success : function(){ // 콜백함수
-        alert("삭제 성공");
-        showList(bno);
-      },
-      error   : function(){ alert("삭제 실패") } // 에러가 발생했을 때, 호출될 함수
-    }); // $.ajax()
-    //alert("삭제 작동");
+    let cno = $(this).closest("li").attr("data-cno");
+    let bno = $(this).closest("li").attr("data-bno");
+    if (confirm("정말 삭제하시겠습니까?")){
+      $.ajax({
+        type:'delete',       // 요청 메서드
+        url: '/instorage/comments/'+cno+'?bno='+bno,  // 요청 URI
+        dataType : 'json', // 전송받을 데이터의 타입
+        success : function(data){ // 콜백함수
+          $("#error").css("display","flex");
+          $("#error p:first-child").html(data.res);
+          showList(bno);
+        },
+        error   : function(data){
+          $("#error").css("display","flex");
+          $("#error p:first-child").html(data.responseJSON.res); } // 에러가 발생했을 때, 호출될 함수
+      }); // $.ajax()
+    }
+
   })
   // *********************************************************************************************************
   // 답글을 만들어준다
@@ -217,7 +232,12 @@
     // 보이게 만들고
     $("#replyBox").css("display","block");
     // 답글 버튼 부모의 자식으로 만들다.
-    $("#replyBox").appendTo($(this).parent())
+    $("#replyBox").appendTo($(this).closest("li"));
+  })
+  $("#commentContainer").on("click","#reply_cancel_btn",function(){
+    $("#replyBox").css("display","none");
+    $("#replyInput").val('');
+    $("#replyBox").appendTo("body");
   })
 
   // 답글 저장 로직
@@ -228,7 +248,6 @@
   let replyWrite = function (){
     let comment = $("#replyInput").val();
     let cno=$("#replyBox").parent().attr("data-cno");
-    console.log(cno);
     if(comment.trim() == ''){
       alert("내용을 입력해주세요.");
       $("#replInput").focus();
@@ -237,50 +256,75 @@
     $.ajax({
       type:'post',       // 요청 메서드
       url: '/instorage/comments?bno='+bno,  // 요청 URI
-      dataType : 'text', // 전송받을 데이터의 타입
+      dataType : 'json', // 전송받을 데이터의 타입
       headers : { "content-type": "application/json"}, // 요청 헤더
       data : JSON.stringify({"pcno":cno,"comment":comment}),
-      success : function(result){ // 콜백함수
-        alert("댓글 등록 성공");
+      success : function(data){ // 콜백함수
         showList(bno);
       },
-      error   : function(){ alert("error") } // 에러가 발생했을 때, 호출될 함수
+      error   : function(data){
+        $("#error").css("display","flex");
+        $("#error p:first-child").html(data.responseJSON.res); } // 에러가 발생했을 때, 호출될 함수
     }); // $.ajax()
     $("#replyBox").css("display","none");
     $("#replyInput").val('');
     $("#replyBox").appendTo("body");
   }
   // ****************************************************************************
+  function dateFormatter(reg){
+    let date = new Date(reg);
 
+    let year = date.getFullYear();
+    let month = (date.getMonth() + 1) < 10 ? '0' + (date.getMonth() + 1) : (date.getMonth() + 1);
+    let day = date.getDate() < 10 ? '0' + date.getDate() : date.getDate();
+
+    let hours = date.getHours() < 10 ? '0' + date.getHours() : date.getHours();
+    let minutes = date.getMinutes() < 10 ? '0' + date.getMinutes() : date.getMinutes();
+    let seconds = date.getSeconds() < 10 ? '0' + date.getSeconds() : date.getSeconds();
+
+    let formattedDate = year + '-' + month + '-' + day + ' ' + hours + ':' + minutes + ':' + seconds;
+    return formattedDate;
+  }
 
 
   let doHtml = function (comments){
     let tmp = "<ul>";
     comments.forEach(function (comments){
-      tmp = tmp + "<li class='comment_li' data-cno='"+comments.cno+"'"
+      let date = dateFormatter(comments.reg_date);
+      tmp = tmp + "<li data-cno='"+comments.cno+"'"
       tmp = tmp + " data-bno='"+comments.bno+"'"
-      tmp = tmp + " data-pcno='"+comments.pcno+"'>"
+      tmp = tmp + " data-pcno='"+comments.pcno+"'"
       if(comments.cno != comments.pcno){
-        tmp = tmp + "ㄴ"
+        tmp = tmp + "class='comment_liR' >"
+      }else{
+        tmp = tmp + "class='comment_li' >"
       }
-      tmp = tmp +"<div class='comment_li_commenter'>"
-      tmp = tmp +"<div class='comment_li_commenter_commenter'>"
-      tmp = tmp + " commenter= "+comments.commenter
-      tmp = tmp + "</div>"
-      tmp = tmp + "<div class = 'comment_li_commenter_update'>"
-      tmp = tmp + "<button class='updateBtn'>수정</button>"
-      tmp = tmp + "<button class='deleteBtn'>삭제</button>"
-      tmp = tmp + "</div>"
-      tmp = tmp + "</div>"
-      tmp = tmp + "<div class='comment_li_comment'>"
-      tmp = tmp +"<span>"+comments.comment+"</span>"
-      tmp = tmp + "</div>"
-      tmp = tmp + "<div class='comment_li_reply'>"
-      if(comments.cno == comments.pcno)
-        tmp = tmp + "<button class='replyBtn'>답글</button>"
-      tmp = tmp + " reg_date= "+comments.reg_date
-      tmp = tmp + "</div>"
-      tmp = tmp + "</li>"
+      if(comments.deleted == 1){
+        tmp = tmp +"<div class='comment_li_commenter'>"
+        tmp = tmp +"<div class='comment_li_commenter_commenter'>"
+        tmp = tmp + ""+comments.commenter
+        tmp = tmp + "</div>"
+        tmp = tmp + "<div class = 'comment_li_commenter_update'>"
+        tmp = tmp + "<button class='updateBtn'>수정</button>"
+        tmp = tmp + "<button class='deleteBtn'>삭제</button>"
+        tmp = tmp + "</div>"
+        tmp = tmp + "</div>"
+        tmp = tmp + "<div class='comment_li_comment'>"
+        tmp = tmp +"<span>"+comments.comment+"</span>"
+        tmp = tmp + "</div>"
+        tmp = tmp + "<div class='comment_li_reply'>"
+        if(comments.cno == comments.pcno)
+          tmp = tmp + "<button class='replyBtn'>답글</button>"
+        else{
+          tmp = tmp + "<div id='tmp'></div>"
+        }
+        tmp = tmp + "<div id='comment_li_reg_date'>"+date+"</div>"
+        tmp = tmp + "</div>"
+        tmp = tmp + "</li>"
+      }else{
+        tmp = tmp +"<div id='deleted_comment'>삭제된 댓글 입니다.</div>"
+      }
+
     })
     return tmp+"</ul>";
   }
